CSS pour page Mon Compte ok

+++ groupomania/src/pages/MonCompte.js
@@ -44,28 +44,47 @@ function MonCompte() {
     }
 
     const deleteUser = () => {
-        deleteRequest(id, articleId, item)
-        history.push('/')
-        localStorage.clear()
+        const resultat = window.confirm("Cette action surpprimera votre compte ainsi que toutes vos pubilications, commentaires et likes. Êtes-vous sur de vouloir continuer ?")
+        if (resultat) {
+            deleteRequest(id, articleId, item)
+            history.push('/')
+            localStorage.clear()
+        } else {
+            return
+        }
     }
 
     return isOpen ? (
-        <form className='monCompte'>
-            <label> Nouveau Username : </label>
-            <input type='text' placeholder="username" required onChange={(e) => {setData({ ...data, username: e.target.value })}}/> 
-            <label> Nouvelle adresse mail : </label>
-            <input type='email' placeholder="@mail" required onChange={(e) => {setData({ ...data, email: e.target.value })}}/> 
-            <label> Nouveau mot de passe : </label>
-            <input type='password' placeholder="********" required onChange={(e) => {setData({ ...data, password: e.target.value })}}/> 
-            <button onClick={submit}> Enregistrer les modifications </button>
-        </form>
+        <div className="monCompte">
+            <h1> Modifier mon compte : </h1> 
+            <form className='monCompte-form'>
+                <div className="monCompte-input">
+                    <label className='monCompte-label'> Nouveau Username : </label>
+                    <input type='text' placeholder="username" required onChange={(e) => {setData({ ...data, username: e.target.value })}}/>
+                </div>
+                <div className="monCompte-input">
+                    <label className='monCompte-label'> Nouvelle adresse mail : </label>
+                    <input type='email' placeholder="@mail" required onChange={(e) => {setData({ ...data, email: e.target.value })}}/> 
+                </div> 
+                <div className="monCompte-input">
+                    <label className='monCompte-label'> Nouveau mot de passe : </label>
+                    <input type='password' placeholder="********" required onChange={(e) => {setData({ ...data, password: e.target.value })}}/>
+                </div> 
+                <button className='monCompte-form-btn'  onClick={submit}> Enregistrer </button>
+            </form>
+        </div>
     ) : (
         <div className='monCompte'>
-            <p> Username : {username} </p>
-            <p> Adresse mail : {email} </p>
-            <p> Mot de passe : ******** </p>
-            <button className="btn" onClick={modify}> 🖊️ Modifier  </button>
-            <button className="btn btn-alert" onClick={deleteUser}> 🗑️ Supprimer le compte </button>
+            <h1> Mon Compte : </h1>
+            <div className="monCompte-info">
+                <p> Username : {username} </p>
+                <p> Adresse mail : {email} </p>
+                <p> Mot de passe : ******** </p>
+            </div>
+            <div className="monCompte-btn">
+                <button className="btn" onClick={modify}> 🖊️ Modifier  </button>
+                <button className="btn btn-alert" onClick={deleteUser}> 🗑️ Supprimer le compte </button>
+            </div>
         </div>
     )
 }
